% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortfunctions.R
\name{sortfunctions}
\alias{sortfunctions}
\title{Insert Global Environment Functions in Current Selection}
\usage{
sortfunctions()
}
\description{
Insert Global Environment Functions in Current Selection
}
